+++ packages/patrol_cli/lib/src/features/drive/constants.dart
@@ -3,3 +3,4 @@ const envPortKey = 'PATROL_PORT';
 const envVerboseKey = 'PATROL_VERBOSE';
 const envPackageNameKey = 'PATROL_APP_PACKAGE_NAME';
 const envBundleIdKey = 'PATROL_APP_BUNDLE_ID';
+const envWaitKey = 'PATROL_WAIT';