+++ packages/patrol_cli/CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Add new `--wait` argument which accepts the number of seconds to wait after
   the test finishes (#251)
+- Make `maestro drive` run all tests (#253)
 
 ## 0.5.2